Formatting and Quality of Life Changes

+++ databasetkinter.py
@@ -1,8 +1,20 @@
 from tkinter import *
 import sqlite3
+import time
 
 root = Tk()
 root.title("Database App")
+
+def validate_input(new_text):
+    if not new_text:
+        return True
+    try:
+        int(new_text)
+        return True
+    except ValueError:
+        return False
+    
+validation = root.register(validate_input)
 
 '''
 d.execute(""" CREATE TABLE students (
@@ -13,17 +25,25 @@ d.execute(""" CREATE TABLE students (
               )
           """)
 '''
-
+    
 def delete():
     data = sqlite3.connect('school.db')
     d = data.cursor()
-    
-    d.execute("DELETE from students WHERE oid = " + delete_ent.get())
-    delete_ent.delete(0,END)
-    
+    d.execute(f"SELECT oid FROM students")
+    oids = d.fetchall()
+    oid_integers = [int(oid[0]) for oid in oids]
+    if (int(delete_ent.get()) >= oid_integers[0]) and (int(delete_ent.get()) <= oid_integers[len(oid_integers)-1]):
+        d.execute("DELETE from students WHERE oid = " + delete_ent.get())
+        delete_ent.delete(0,END)
+    else:
+        warning_label= Label(root,text="Invalid ID", width=30, anchor="center")
+        warning_label.grid(row=7,column=0,columnspan=2)
     data.commit()
     data.close()
+    
+
 def submit():
+    global window
     if (int(grade.get()) >= 1 and int(grade.get()) <= 10) and (len(phone_no.get()) == 11):
         data = sqlite3.connect('school.db')
         d = data.cursor()
@@ -34,23 +54,30 @@ def submit():
                    "class":grade.get(),
                    "phone_no":phone_no.get()}
         )
-        success_label= Label(root,text=str(firstname.get()) +" "+str(lastname.get())+" added!")
+        success_label= Label(root,text=str(firstname.get()) +" "+str(lastname.get())+" added!", width=30, anchor="center")
         success_label.grid(row=7,column=0,columnspan=2)
         firstname.delete(0,END)
         lastname.delete(0,END)
         grade.delete(0,END)
         phone_no.delete(0,END)
-        
         data.commit()
         data.close()
+        
+        try:
+            window.destroy()
+        except NameError:
+            pass
+        finally:
+            show()
     else:
         if not((int(grade.get()) >= 1 and int(grade.get()) <= 10)) and not(len(phone_no.get()) == 11):
-            warning_label= Label(root,text="Invalid Class and Phone Number")
+            warning_label= Label(root,text="Invalid Class and Phone Number", width=30, anchor="center")
         elif not(len(phone_no.get()) == 11):
-            warning_label= Label(root,text="Invalid Phone Number")
+            warning_label= Label(root,text="Invalid Phone Number", width=30, anchor="center")
         elif not((int(grade.get()) >= 1 and int(grade.get()) <= 10)):
-            warning_label= Label(root,text="Invalid Class")
+            warning_label= Label(root,text="Invalid Class", width=30, anchor="center")
         warning_label.grid(row=7,column=0,columnspan=2)
+   
     
 def show():
     global window
@@ -108,7 +135,7 @@ def show():
         i += 1
         
     update_btn = Button(window,text = "Update Record",command = update)
-    update_btn.grid(row = i,column=2,pady= (10,5))
+    update_btn.grid(row = i,column=2,pady = (10,5))
         
     data.commit()
     data.close()
@@ -155,17 +182,17 @@ def edit():
     lastname_editor = Entry(editor,width=20)
     lastname_editor.grid(row=1, column=1)
 
-    grade_editor = Entry(editor,width=20)
+    grade_editor = Entry(editor,width=20,validate="key", validatecommand=(validation, '%P'))
     grade_editor.grid(row=2, column=1)
 
-    phone_no_editor = Entry(editor,width=20)
+    phone_no_editor = Entry(editor,width=20,validate="key", validatecommand=(validation, '%P'))
     phone_no_editor.grid(row=3, column=1)
     
     for record in records:
         firstname_editor.insert(0,record[0])
         lastname_editor.insert(0,record[1])
         grade_editor.insert(0,record[2])
-        phone_no_editor.insert(0,record[3])
+        phone_no_editor.insert(0,"0"+str(record[3]))
 
     #Creating Label
     firstname_label = Label(editor, text= "First Name" )
@@ -192,7 +219,7 @@ def update():
     input_window = Toplevel()
     prompt = Label(input_window, text = "Enter ID which you want to edit")
     prompt.pack()
-    id_input = Entry(input_window,width = 15,font = ("Helvetica",18,"bold"))
+    id_input = Entry(input_window,width = 15,font = ("Helvetica",18,"bold"),validate="key", validatecommand=(validation, '%P'))
     id_input.pack()
     enter = Button(input_window,text = "Edit",font = ("Helvetica",12),width=10,command = edit)
     enter.pack(pady= 10)
@@ -205,13 +232,13 @@ firstname.grid(row=0, column=1)
 lastname = Entry(root,width=20)
 lastname.grid(row=1, column=1)
 
-grade = Entry(root,width=20)
+grade = Entry(root,width=20,validate="key", validatecommand=(validation, '%P'))
 grade.grid(row=2, column=1)
 
-phone_no = Entry(root,width=20)
+phone_no = Entry(root,width=20,validate="key", validatecommand=(validation, '%P'))
 phone_no.grid(row=3, column=1)
 
-delete_ent = Entry(root,width=20)
+delete_ent = Entry(root,width=20,validate="key", validatecommand=(validation, '%P'))
 delete_ent.grid(row=5, column=1)
 
 #Creating Label
@@ -235,7 +262,7 @@ submit_btn.grid(row=4,column=0,pady=15)
 show_btn = Button(root, text = "Show Records",command=show,padx=20)
 show_btn.grid(row=4,column=1,pady=5)
 delete_btn = Button(root, text = "Delete",command=delete,padx=20,width=28)
-delete_btn.grid(row=6, column=0,columnspan=2)
+delete_btn.grid(row=6, column=0,columnspan=2,padx = 5,pady=5)
